chore: Remove proguard plugin to simplify build configuration
Due to the change in current project requirements, it is no longer necessary to use the proguard plugin for code obfuscation, so the proguard plugin configuration is removed from the build.gradle.kts file. This change is intended to simplify build configuration, reduce unnecessary plugin dependencies, and optimize the build process.